Modified code to work based on Input Capture. Changed output data type to double because we are outputting real world time now.

+++ include/ultrasonic.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void setup_ultrasonic(void);
-uint32_t read_ultrasonic(void);
+double read_ultrasonic(void);
 
 #endif // ULTRASONIC_H
